+++ MetaBond.Application/DTOs/Communities/CommunitiesDTos.cs
@@ -1,10 +1,9 @@
-﻿namespace MetaBond.Application.DTOs.Communities
-{
-    public sealed record CommunitiesDTos
-    (
-        Guid CommunitiesId,
-        string? Name,
-        string? Category,
-        DateTime CreatedAt
-    );
-}
+﻿namespace MetaBond.Application.DTOs.Communities;
+
+public sealed record CommunitiesDTos
+(
+    Guid CommunitiesId,
+    string? Name,
+    string? Category,
+    DateTime CreatedAt
+);
