+++ MetaBond.Application/DTOs/Communities/PostsAndEventsDTos.cs
@@ -1,12 +1,11 @@
-﻿namespace MetaBond.Application.DTOs.Communities
-{
-    public sealed record PostsAndEventsDTos
-    (
-        Guid CommunitieId,
-        string? Name,
-        string? Category,
-        DateTime CreatedAt,
-        IEnumerable<Domain.Models.Posts> Posts,
-        IEnumerable<Domain.Models.Events> Events
-    );
-}
+﻿namespace MetaBond.Application.DTOs.Communities;
+
+public sealed record PostsAndEventsDTos
+(
+    Guid CommunitiesId,
+    string? Name,
+    string? Category,
+    DateTime CreatedAt,
+    IEnumerable<Domain.Models.Posts> Posts,
+    IEnumerable<Domain.Models.Events> Events
+);
