+++ MetaBond.Application/DTOs/Events/CommunitiesDTos.cs
@@ -1,13 +1,10 @@
-﻿using MetaBond.Domain.Models;
-namespace MetaBond.Application.DTOs.Events
-{
-    public sealed record CommunitiesDTos
-    (
-        Guid? Id,
-        string? Description,
-        string? Title,
-        DateTime? DateAndTime,
-        DateTime? CreatedAt,
-        List<CommunitySummaryDto> Communities
-    );
-}
+﻿namespace MetaBond.Application.DTOs.Events;
+public sealed record CommunitiesDTos
+(
+    Guid? Id,
+    string? Description,
+    string? Title,
+    DateTime? DateAndTime,
+    DateTime? CreatedAt,
+    List<CommunitySummaryDto> Communities
+);
