+++ MetaBond.Application/DTOs/Events/CommunitySummaryDto.cs
@@ -1,7 +1,8 @@
 namespace MetaBond.Application.DTOs.Events;
 
-public record CommunitySummaryDto(
+public sealed record CommunitySummaryDto
+(
     string? Description,
     string? Category,
     DateTime CreatedAt
-    );
+);
