+++ MetaBond.Application/DTOs/Events/EventsDto.cs
@@ -1,13 +1,12 @@
 ﻿
-namespace MetaBond.Application.DTOs.Events
-{
-    public sealed record EventsDto
-    (
-        Guid? Id,
-        string? Description,
-        string? Title,
-        DateTime? DateAndTime,
-        DateTime? CreatedAt,
-        Guid? CommunitiesId
-    );
-}
+namespace MetaBond.Application.DTOs.Events;
+
+public sealed record EventsDto
+(
+    Guid? Id,
+    string? Description,
+    string? Title,
+    DateTime? DateAndTime,
+    DateTime? CreatedAt,
+    Guid? CommunitiesId
+);
