+++ MetaBond.Application/DTOs/Events/EventsWithParticipationInEventsDTos.cs
@@ -1,14 +1,11 @@
-﻿using MetaBond.Domain.Models;
+﻿namespace MetaBond.Application.DTOs.Events;
 
-namespace MetaBond.Application.DTOs.Events
-{
-    public sealed record EventsWithParticipationInEventsDTos
-    (
-        Guid? EventsId,
-        string? Description,
-        string? Title,
-        DateTime? DateAndTime,
-        IEnumerable<ParticipationInEventBasicDTos> ParticipationInEvents,
-        DateTime? CreatedAt
-    );
-}
+public sealed record EventsWithParticipationInEventsDTos
+(
+    Guid? EventsId,
+    string? Description,
+    string? Title,
+    DateTime? DateAndTime,
+    IEnumerable<ParticipationInEventBasicDTos> ParticipationInEvents,
+    DateTime? CreatedAt
+);
