+++ MetaBond.Application/DTOs/Friendship/FriendshipDTos.cs
@@ -1,13 +1,12 @@
 ﻿using MetaBond.Domain;
 
-namespace MetaBond.Application.DTOs.Friendship
-{
-    public sealed record FriendshipDTos
-    (
-        Guid FriendshipId,
-        Status Status,
-        Guid? RequesterId,
-        Guid? AddresseeId,
-        DateTime? CreatedAt
-    );
-}
+namespace MetaBond.Application.DTOs.Friendship;
+
+public sealed record FriendshipDTos
+(
+    Guid FriendshipId,
+    Status Status,
+    Guid? RequesterId,
+    Guid? AddresseeId,
+    DateTime? CreatedAt
+);
