+++ MetaBond.Application/DTOs/ParticipationInEventDtos/ParticipationInEventDTos.cs
@@ -1,9 +1,8 @@
 ﻿
-namespace MetaBond.Application.DTOs.ParticipationInEvent
-{
-    public sealed record ParticipationInEventDTos
-    (
-        Guid? ParticipationInEventId,
-        Guid? EventId
-    );
-}
+namespace MetaBond.Application.DTOs.ParticipationInEventDtos;
+
+public sealed record ParticipationInEventDTos
+(
+    Guid? ParticipationInEventId,
+    Guid? EventId
+);
