+++ MetaBond.Application/DTOs/ProgressEntry/ProgressBoardSummaryDTos.cs
@@ -1,6 +1,6 @@
 namespace MetaBond.Application.DTOs.ProgressEntry;
 
-public record ProgressBoardSummaryDTos
+public sealed record ProgressBoardSummaryDTos
 (
    Guid ProgressBoardId,
    Guid UserId,
