+++ MetaBond.Application/DTOs/ProgressEntry/ProgressEntryBasicDTos .cs
@@ -1,5 +1,4 @@
 ﻿
-namespace MetaBond.Application.DTOs.ProgressEntry
-{
-    public sealed record ProgressEntryBasicDTos(Guid ProgressEntryId, string? Description, Guid ProgressBoardId,Guid UserId);
-}
+namespace MetaBond.Application.DTOs.ProgressEntry;
+
+public sealed record ProgressEntryBasicDTos(Guid ProgressEntryId, string? Description, Guid ProgressBoardId,Guid UserId);
