+++ MetaBond.Application/DTOs/ProgressEntry/ProgressEntryDTos.cs
@@ -1,13 +1,12 @@
 ﻿
-namespace MetaBond.Application.DTOs.ProgressEntry
-{
-    public sealed record ProgressEntryDTos
-    (
-        Guid ProgressEntryId,
-        Guid ProgressBoardId,
-        Guid UserId,
-        string? Description,
-        DateTime CreatedAt,
-        DateTime UpdateAt
-    );
-}
+namespace MetaBond.Application.DTOs.ProgressEntry;
+
+public sealed record ProgressEntryDTos
+(
+    Guid ProgressEntryId,
+    Guid ProgressBoardId,
+    Guid UserId,
+    string? Description,
+    DateTime CreatedAt,
+    DateTime UpdateAt
+);
