+++ MetaBond.Application/DTOs/ProgressEntry/ProgressEntryWithBoardDTos.cs
@@ -2,7 +2,7 @@ using MetaBond.Application.DTOs.ProgressBoard;
 
 namespace MetaBond.Application.DTOs.ProgressEntry;
 
-public record ProgressEntryWithBoardDTos
+public sealed record ProgressEntryWithBoardDTos
 (
     Guid ProgressEntryId, 
     string? Description,
