+++ MetaBond.Application/DTOs/ProgressEntry/ProgressEntryWithDescriptionDTos.cs
@@ -1,5 +1,3 @@
 ﻿
-namespace MetaBond.Application.DTOs.ProgressEntry
-{
-    public sealed record ProgressEntryWithDescriptionDTos(string? Description);
-}
+namespace MetaBond.Application.DTOs.ProgressEntry;
+public sealed record ProgressEntryWithDescriptionDTos(string? Description);
