+++ MetaBond.Application/DTOs/ProgressEntry/ProgressEntryWithProgressBoardDTos.cs
@@ -2,7 +2,7 @@ using MetaBond.Application.DTOs.ProgressBoard;
 
 namespace MetaBond.Application.DTOs.ProgressEntry;
 
-public record ProgressEntryWithProgressBoardDTos
+public sealed record ProgressEntryWithProgressBoardDTos
 (
     Guid ProgressEntryId,
     Guid UserId,
