+++ MetaBond.Application/DTOs/Rewards/RewardsDTos.cs
@@ -1,12 +1,10 @@
 ﻿
-namespace MetaBond.Application.DTOs.Rewards
-{
-    public sealed record RewardsDTos
-    (
-        Guid? RewardsId,
-        Guid UserId,
-        string? Description,
-        int? PointAwarded,
-        DateTime? DateAwarded
-    );
-}
+namespace MetaBond.Application.DTOs.Rewards;
+public sealed record RewardsDTos
+(
+    Guid? RewardsId,
+    Guid UserId,
+    string? Description,
+    int? PointAwarded,
+    DateTime? DateAwarded
+);
